Datasets
Information about the dataset:
The data contains two files: train.csv and test.csv

train.csv - which contains the training set observations. 86,845 rows, which have the target variable (attack) values as well for training the model.

test.csv - which contains the testing set observations. 21,712 rows, with the target column missing. Use the trained model to predict and submit the values of the target column.


Additionally, a sample-submission.csv file is provided to show you what your CSV submission should look like. Note that the header has 'attack' for the column name and the number of rows is equal to the number of rows in the test set.

attack: normal - 0

attack: neptune - 1